Translate analogWriteResolution.adoc
with values between 0 and 4095 to exploit the full DAC resolution or to set the PWM signal without rolling over.

Could you please explain this sentence?
I cannot understand the meaning of "without rolling over".

And, I will go to UK University. So, I was busy for preparing and I couldn't translate. Sorry!